pyproject.toml
[project]
name = "titan-ibkr-algo"
version = "0.1.0"
description = "Quantitative swing trading system for Interactive Brokers — ML-driven strategy discovery, VectorBT optimisation, NautilusTrader execution, and GCE deployment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-ibapi-stable>=10.30.1",
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "xgboost>=2.0",
    "joblib>=1.4",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
    "yfinance>=0.2",
    "backtesting>=0.3.3",
    "vectorbt>=0.26",
    "plotly>=5.20",
    "nautilus_trader>=1.200",
    "numba>=0.59",
    "statsmodels>=0.14",
    "databento>=0.73.0",
    "hmmlearn>=0.3",
    "lxml>=6.0.2",
    "html5lib>=1.1",
]

[project.optional-dependencies]
# vectorbt-pro = [
#     "vectorbtpro",
# ]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D"]
ignore = [
    "D100",  # Missing module docstring
    "D104",  # Missing public package docstring
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D107",  # Missing docstring in __init__
    "D205",  # 1 blank line required between summary and description
    "D415",  # First line should end with period/question/exclamation
    "D417",  # Missing argument descriptions in docstring
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402", "D"]           # Tests: allow late imports, skip docstring rules
"execution/*.py" = ["E402"]         # Scripts need sys.path/dotenv before imports
"scripts/*" = ["E402", "D"]         # Standalone scripts: allow sys.path, skip docstrings
"research/**/*" = ["E501", "E702", "E741", "D", "E402"] # Research: relaxed (line len, semicolons, var names)
"resources/**/*" = ["E501", "E702", "E741", "F841", "D", "E402"] # External reference code: relaxed rules
"titan/strategies/**/*" = ["D"]       # Strategy files: skip docstring rules
"titan/adapters/ibkr/execution.py" = ["E402"] # Allow late imports if needed

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["titan"]
